Refactor - moved OpponentsPoisonCountersCount to shared area
